+++ src/main/java/com/andersmmg/cityessentials/block/custom/SpeedLimitSignBlock.java
@@ -1,0 +1,122 @@
+package com.andersmmg.cityessentials.block.custom;
+
+import com.andersmmg.cityessentials.block.entity.SpeedLimitSignBlockEntity;
+import com.andersmmg.cityessentials.client.screen.SpeedLimitEditScreen;
+import com.andersmmg.cityessentials.item.ModItems;
+import com.andersmmg.cityessentials.util.VoxelUtils;
+import net.fabricmc.api.EnvType;
+import net.fabricmc.api.Environment;
+import net.minecraft.block.*;
+import net.minecraft.block.entity.BlockEntity;
+import net.minecraft.client.MinecraftClient;
+import net.minecraft.entity.player.PlayerEntity;
+import net.minecraft.inventory.Inventory;
+import net.minecraft.item.ItemPlacementContext;
+import net.minecraft.server.world.ServerWorld;
+import net.minecraft.state.StateManager;
+import net.minecraft.state.property.DirectionProperty;
+import net.minecraft.state.property.Properties;
+import net.minecraft.util.*;
+import net.minecraft.util.function.BooleanBiFunction;
+import net.minecraft.util.hit.BlockHitResult;
+import net.minecraft.util.math.BlockPos;
+import net.minecraft.util.math.Direction;
+import net.minecraft.util.math.random.Random;
+import net.minecraft.util.shape.VoxelShape;
+import net.minecraft.util.shape.VoxelShapes;
+import net.minecraft.world.BlockView;
+import net.minecraft.world.World;
+import org.jetbrains.annotations.Nullable;
+
+public class SpeedLimitSignBlock extends BlockWithEntity {
+    public static final DirectionProperty FACING = Properties.HORIZONTAL_FACING;
+    private static final VoxelShape VOXEL_SHAPE = VoxelShapes.combineAndSimplify(Block.createCuboidShape(6, 0, 6, 10, 12, 10), Block.createCuboidShape(0, 0, 6, 16, 16, 7), BooleanBiFunction.OR);
+
+    public SpeedLimitSignBlock(Settings settings) {
+        super(settings);
+        this.setDefaultState(this.stateManager.getDefaultState().with(FACING, Direction.NORTH));
+    }
+
+    protected static Direction getDirection(BlockState state) {
+        return state.get(FACING);
+    }
+
+    @Override
+    public ActionResult onUse(BlockState state, World world, BlockPos pos, PlayerEntity player, Hand hand, BlockHitResult hit) {
+        BlockEntity blockEntity = world.getBlockEntity(pos);
+        if (blockEntity instanceof SpeedLimitSignBlockEntity) {
+            if (player.getStackInHand(hand).isOf(ModItems.MARKER)) {
+                if (world.isClient) {
+                    showEditScreen((SpeedLimitSignBlockEntity) blockEntity);
+                }
+            }
+        }
+        return ActionResult.CONSUME;
+    }
+
+    @Environment(EnvType.CLIENT)
+    private void showEditScreen(SpeedLimitSignBlockEntity blockEntity) {
+        if (blockEntity == null) {
+            return;
+        }
+        SpeedLimitEditScreen screen = new SpeedLimitEditScreen(blockEntity);
+        MinecraftClient.getInstance().setScreen(screen);
+    }
+
+    @Override
+    public void onStateReplaced(BlockState state, World world, BlockPos pos, BlockState newState, boolean moved) {
+        if (state.isOf(newState.getBlock())) {
+            return;
+        }
+        BlockEntity blockEntity = world.getBlockEntity(pos);
+        if (blockEntity instanceof Inventory) {
+            ItemScatterer.spawn(world, pos, (Inventory) ((Object) blockEntity));
+            world.updateComparators(pos, this);
+        }
+        super.onStateReplaced(state, world, pos, newState, moved);
+    }
+
+    @Override
+    public void scheduledTick(BlockState state, ServerWorld world, BlockPos pos, Random random) {
+        BlockEntity blockEntity = world.getBlockEntity(pos);
+        if (blockEntity instanceof SpeedLimitSignBlockEntity) {
+            ((SpeedLimitSignBlockEntity) blockEntity).tick();
+        }
+    }
+
+    @Override
+    @Nullable
+    public BlockEntity createBlockEntity(BlockPos pos, BlockState state) {
+        return new SpeedLimitSignBlockEntity(pos, state);
+    }
+
+    @Override
+    public BlockRenderType getRenderType(BlockState state) {
+        return BlockRenderType.MODEL;
+    }
+
+    @Override
+    public BlockState rotate(BlockState state, BlockRotation rotation) {
+        return (BlockState) state.with(FACING, rotation.rotate(state.get(FACING)));
+    }
+
+    @Override
+    public BlockState mirror(BlockState state, BlockMirror mirror) {
+        return state.rotate(mirror.getRotation(state.get(FACING)));
+    }
+
+    @Override
+    public VoxelShape getOutlineShape(BlockState state, BlockView world, BlockPos pos, ShapeContext context) {
+        return VoxelUtils.rotateShape(getDirection(state), VOXEL_SHAPE);
+    }
+
+    @Override
+    protected void appendProperties(StateManager.Builder<Block, BlockState> builder) {
+        builder.add(FACING);
+    }
+
+    @Override
+    public BlockState getPlacementState(ItemPlacementContext ctx) {
+        return (BlockState) this.getDefaultState().with(FACING, ctx.getHorizontalPlayerFacing().getOpposite());
+    }
+}