+++ src/main/java/com/andersmmg/cityessentials/client/renderer/SpeedLimitSignBlockEntityRenderer.java
@@ -1,7 +1,7 @@
 package com.andersmmg.cityessentials.client.renderer;
 
-import com.andersmmg.cityessentials.block.custom.StopSignBlock;
-import com.andersmmg.cityessentials.block.entity.StopSignBlockEntity;
+import com.andersmmg.cityessentials.block.custom.SpeedLimitSignBlock;
+import com.andersmmg.cityessentials.block.entity.SpeedLimitSignBlockEntity;
 import net.minecraft.block.BlockState;
 import net.minecraft.client.MinecraftClient;
 import net.minecraft.client.font.TextRenderer;
@@ -15,17 +15,17 @@ import net.minecraft.text.Text;
 import net.minecraft.util.Colors;
 import net.minecraft.util.math.*;
 
-public class StopSignBlockEntityRenderer implements BlockEntityRenderer<StopSignBlockEntity> {
+public class SpeedLimitSignBlockEntityRenderer implements BlockEntityRenderer<SpeedLimitSignBlockEntity> {
     private final TextRenderer textRenderer;
 
     private static final int RENDER_DISTANCE = MathHelper.square(64);
 
-    public StopSignBlockEntityRenderer(BlockEntityRendererFactory.Context ctx) {
+    public SpeedLimitSignBlockEntityRenderer(BlockEntityRendererFactory.Context ctx) {
         this.textRenderer = ctx.getTextRenderer();
     }
 
     @Override
-    public void render(StopSignBlockEntity entity, float tickDelta, MatrixStack matrices, VertexConsumerProvider vertexConsumers, int light, int overlay) {
+    public void render(SpeedLimitSignBlockEntity entity, float tickDelta, MatrixStack matrices, VertexConsumerProvider vertexConsumers, int light, int overlay) {
         if (!shouldRender(entity.getPos())) {
             return;
         }
@@ -33,16 +33,24 @@ public class StopSignBlockEntityRenderer implements BlockEntityRenderer<StopSign
         matrices.translate(0.5f, 0.5f, 0.5f);
         matrices.multiply(RotationAxis.POSITIVE_Y.rotationDegrees(getRotation(entity)));
         matrices.translate(0.0f, 0.0f, -0.11f);
-        float textScale = 0.030f;
+        float textScale = 0.019f;
+        float speedTextScale = 2.2f;
         matrices.scale(textScale, textScale, textScale);
         matrices.multiply(RotationAxis.POSITIVE_Z.rotationDegrees(180));
 
-        Text signText = Text.literal("STOP");
+        Text firstText = Text.literal("SPEED");
+        Text secondText = Text.literal("LIMIT");
+        Text signText = entity.getText();
 
         float f = (float) (-this.textRenderer.getWidth(signText) / 2);
         float g = (float) (-this.textRenderer.fontHeight / 2);
 
-        this.textRenderer.draw(signText, f + 0.5f, g + 0.5f, Colors.WHITE, false, matrices.peek().getPositionMatrix(), vertexConsumers, TextRenderer.TextLayerType.POLYGON_OFFSET, 0, 255);
+        this.textRenderer.draw(firstText, ((float) -this.textRenderer.getWidth(firstText) / 2), g - 14, Colors.BLACK, false, matrices.peek().getPositionMatrix(), vertexConsumers, TextRenderer.TextLayerType.POLYGON_OFFSET, 0, 255);
+
+        this.textRenderer.draw(secondText, ((float) -this.textRenderer.getWidth(secondText) / 2), g - 4, Colors.BLACK, false, matrices.peek().getPositionMatrix(), vertexConsumers, TextRenderer.TextLayerType.POLYGON_OFFSET, 0, 255);
+
+        matrices.scale(speedTextScale, speedTextScale, 1);
+        this.textRenderer.draw(signText, f, g + 6, Colors.BLACK, false, matrices.peek().getPositionMatrix(), vertexConsumers, TextRenderer.TextLayerType.POLYGON_OFFSET, 0, 255);
         matrices.pop();
     }
 
@@ -58,9 +66,9 @@ public class StopSignBlockEntityRenderer implements BlockEntityRenderer<StopSign
 
     }
 
-    public int getRotation(StopSignBlockEntity entity) {
+    public int getRotation(SpeedLimitSignBlockEntity entity) {
         BlockState state = entity.getCachedState();
-        Direction dir = state.get(StopSignBlock.FACING);
+        Direction dir = state.get(SpeedLimitSignBlock.FACING);
         return switch (dir) {
             case NORTH -> 0;
             case SOUTH -> 180;